+++ src/Models/VisitorsModel.php
@@ -1220,11 +1220,16 @@ class VisitorsModel extends BaseModel
             'term_id'       => '',
         ]);
 
+        $range = DateRange::get('30days');
+
+        $startDate = $range['from'] . ' 00:00:00';
+        $endDate   = date('Y-m-d', strtotime($range['to'] . ' +1 day')) . ' 00:00:00';
+
         $fields = [
             '`visitor`.`last_counter` AS `date`',
             'COUNT(DISTINCT `visitor`.`ID`) AS `visitors`',
             '`visit`.`visit` AS `visits`',
-            'COUNT(DISTINCT CASE WHEN(`visitor`.`referred` NOT LIKE "%%' . Helper::get_domain_name(home_url()) . '%%" AND `visitor`.`referred` <> "") THEN `visitor`.`ID` END) AS `referrers`',
+            'COUNT(DISTINCT CASE WHEN(`visitor`.`referred` <> "") THEN `visitor`.`ID` END) AS `referrers`',
         ];
         if (is_numeric($args['post_id']) || !empty($args['author_id']) || !empty($args['term_id'])) {
             // For single pages/posts/authors/terms
@@ -1236,9 +1241,9 @@ class VisitorsModel extends BaseModel
             // For single pages/posts/authors/terms
             $query->join('visit', ['`visitor`.`last_counter`', '`visit`.`last_counter`']);
         }
-        $query
-            ->whereDate('`visitor`.`last_counter`', $args['date'])
-            ->groupBy('`visitor`.`last_counter`');
+        $query->where('visitor.last_counter', '>=', $startDate)
+            ->where('visitor.last_counter', '<', $endDate)
+            ->groupBy('visitor.last_counter');
 
         $filteredArgs = array_filter($args);
         if (array_intersect(['post_type', 'post_id', 'resource_type', 'author_id', 'taxonomy', 'term_id'], array_keys($filteredArgs))) {